requirements.txt
flask
python-dotenv
requests
google-genai
pytest